backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
pydantic==2.*
google-cloud-firestore==2.*